+++ changelog.md
@@ -8,6 +8,7 @@ Codion Change Log
 - AbstractScenario, run counting removed.
 - LoadTest.Scenario.maximumTime() removed.
 - LoadTest.Scenario.Builder introduced along with LoadTest.Performer, AbstractScenario removed, related refactoring.
+- LoadTest.Scenario, couple of scenario() factory methods added.
 ### is.codion.common.rmi
 - SerializationWhitelist, now allows arrays but checks the component type.
 - SerializationWhitelist.DryRun, now excludes arrays and includes the component type.